+++ CHANGELOG.md
@@ -11,7 +11,6 @@
          • More features for both widgets.
 
 ## [1.0.1] - Release Date: 20.05.2021
-
    # Widgets:
      • ExpandableWidget - renamed.
      • ExpandableWidget.extended - renamed.
@@ -22,7 +21,6 @@
      • More features for ExpandableTextWidget
 
 ## [1.0.2] - Release Date: 06.06.2021
-
    # Widgets:
      • Expandable & ExpandableText - renamed.
      • Arrow locations for Expandable widgets.
@@ -31,3 +29,15 @@
    # TODOs:
      • Custom animations.
      • More features.
+
+## [1.0.3] - Release Date: TODO
+    # Widgets:
+      • New enum - Clickable
+      • New property - wrapContent
+      • shape property removed -> borderRadius property added.
+      • direction property added.
+      • animation and animationController added for custom animations(sorry for being too late).
+      • Various bug fixes and performance improvements.
+
+    # TODOs:
+      • More features.